+++ internal/colors/distance.go
@@ -5,9 +5,9 @@ import (
 )
 
 func DistanceLab(a, b *Lab) float64 {
-	deltaL := float64(b.L - a.L)
-	deltaA := float64(b.A - a.A)
-	deltaB := float64(b.B - a.B)
+	deltaL := b.L - a.L
+	deltaA := b.A - a.A
+	deltaB := b.B - a.B
 
 	return math.Sqrt(deltaL*deltaL + deltaA*deltaA + deltaB*deltaB)
 }
