+++ internal/colors/rgb.go
@@ -46,7 +46,7 @@ func NRGBtoRGB(in *NRgb) *Rgb {
 }
 
 func RGBtoXYZ(in *Rgb) (*Xyz, error) {
-	workingMatData := [][]float32{
+	workingMatData := [][]float64{
 		{0.4124564, 0.3575761, 0.1804375},
 		{0.2126729, 0.7151522, 0.0721750},
 		{0.0193339, 0.1191920, 0.9503041},
@@ -80,7 +80,7 @@ func RGBtoLab(in *Rgb) (*Lab, error) {
 	}
 
 	// reference white D65
-	var wX, wY, wZ float32
+	var wX, wY, wZ float64
 	wX, wY, wZ = 0.95047003, 1.0000001, 1.08883
 
 	xTrans := labTransform(xyz.X / wX)
@@ -96,9 +96,9 @@ func RGBtoLab(in *Rgb) (*Lab, error) {
 	return lab, nil
 }
 
-func labTransform(t float32) float32 {
+func labTransform(t float64) float64 {
 	if t > 0.008856 {
-		return float32(math.Pow(float64(t), 1.0/3.0))
+		return math.Pow(t, 1.0/3.0)
 	}
 
 	return t*7.787 + 16.0/116.0
@@ -114,11 +114,11 @@ func GammaRemoveNRGB(in *NRgb) *NRgb {
 	return out
 }
 
-func normalizeRGB(value uint8) float32 {
-	return float32(value) / 255.0
+func normalizeRGB(value uint8) float64 {
+	return float64(value) / 255.0
 }
 
-func scaleNRGB(value float32) uint8 {
+func scaleNRGB(value float64) uint8 {
 	if value < 0 {
 		value = 0
 	}
@@ -128,10 +128,10 @@ func scaleNRGB(value float32) uint8 {
 	return uint8(value * 255.0)
 }
 
-func gammaRemove(value float32) float32 {
+func gammaRemove(value float64) float64 {
 	if value <= 0.04045 {
 		return value / 12.92
 	}
 
-	return float32(math.Pow(float64((value+0.055)/1.055), 2.4))
+	return math.Pow((value + 0.055/1.055), 2.4)
 }
