+++ internal/palette/contrasts.go
@@ -7,8 +7,7 @@ import (
 	"github.com/mermonia/chromatika/internal/utils"
 )
 
-
-func wcag(a, b *colors.LCHab) (float32, error) {
+func wcag(a, b *colors.LCHab) (float64, error) {
 	labA := colors.LCHtoLab(a)
 	labB := colors.LCHtoLab(b)
 
@@ -27,20 +26,20 @@ func deltaE00(a, b *colors.LCHab) float64 {
 	labB := colors.LCHtoLab(b)
 
 	// We need a components to create new intermediate colors
-	aA := float64(labA.A)
-	bA := float64(labB.A)
+	aA := labA.A
+	bA := labB.A
 
 	// Average chroma and luminosity
-	avgL := float64((a.L + b.L) / 2)
-	avgC := float64((a.C + b.C) / 2)
+	avgL := (a.L + b.L) / 2
+	avgC := (a.C + b.C) / 2
 
 	// G parameter for future calculations
 	avgCto7 := math.Pow(avgC, 7)
 	G := 0.5 * (1 - math.Sqrt(avgCto7/(avgCto7+6103515625)))
 
 	// Intermediate color generation
-	aAprime := float32(aA * (1 + G))
-	bAprime := float32(bA * (1 + G))
+	aAprime := aA * (1 + G)
+	bAprime := bA * (1 + G)
 
 	labAprime := &colors.Lab{
 		L: a.L,
@@ -58,8 +57,8 @@ func deltaE00(a, b *colors.LCHab) float64 {
 	bPrime := colors.LabToLCH(labBprime)
 
 	// T parameter for future calculations
-	avgHprime := float64(aPrime.H + bPrime.H)
-	if math.Abs(float64(aPrime.H-bPrime.H)) > 180 {
+	avgHprime := aPrime.H + bPrime.H
+	if math.Abs(aPrime.H-bPrime.H) > 180 {
 		avgHprime += 360
 	}
 	avgHprime /= 2
@@ -67,10 +66,10 @@ func deltaE00(a, b *colors.LCHab) float64 {
 	T := 1 - 0.17*utils.DegCos(avgHprime-30) + 0.24*utils.DegCos(2*avgHprime) + 0.32*utils.DegCos(3*avgHprime+6) - 0.20*utils.DegCos(4*avgHprime-63)
 
 	// Delta calculations
-	deltaL := float64(bPrime.L - aPrime.L)
-	deltaC := float64(bPrime.C - aPrime.C)
+	deltaL := bPrime.L - aPrime.L
+	deltaC := bPrime.C - aPrime.C
 
-	hPrimeDiff := float64(bPrime.H - aPrime.H)
+	hPrimeDiff := bPrime.H - aPrime.H
 	var imDeltaH float64
 	if math.Abs(hPrimeDiff) <= 180 {
 		imDeltaH = hPrimeDiff
@@ -80,10 +79,10 @@ func deltaE00(a, b *colors.LCHab) float64 {
 		imDeltaH = hPrimeDiff - 360
 	}
 
-	deltaH := 2 * math.Sqrt(float64(aPrime.C*bPrime.C)) * utils.DegSin(imDeltaH/2)
+	deltaH := 2 * math.Sqrt(aPrime.C*bPrime.C) * utils.DegSin(imDeltaH/2)
 
 	// Calculation of S values
-	avgCprime := float64(aPrime.C+bPrime.C) / 2
+	avgCprime := aPrime.C + bPrime.C/2
 
 	sL := 1 + (0.015*math.Pow(avgL-50, 2))/math.Sqrt(20+math.Pow(avgL-50, 2))
 	sC := 1 + 0.045*avgCprime
